animal APIs and GET requests are in separate js files now
