+++ routes/apiRoutes/index.js
@@ -1,0 +1,8 @@
+const router = require('express').Router();
+const animalRoutes = require('../apiRoutes/animalRoutes');
+
+router.use(animalRoutes);
+
+module.exports = router;
+
+//this file is a central hub for all routing functions
